homework1_requirements.txt
Package                  Version
------------------------ ----------
dnspython                1.16.0
